+++ changelog.md
@@ -25,3 +25,4 @@
 
 * Added `eachParallel` method
 * Added `eachSeries` alias of `each`
+* Tests for execution order for parallel functions